+++ src/flask_marshmallow/compat.py
@@ -5,5 +5,5 @@ This module should be considered private API.
 import marshmallow
 
 _MARSHMALLOW_VERSION_INFO = tuple(
-    [int(part) for part in marshmallow.__version__.split(".") if part.isdigit()]
+    int(part) for part in marshmallow.__version__.split(".") if part.isdigit()
 )